Code cleanup to reduce duplication
