+++ api/app/.env.example
@@ -46,6 +46,7 @@ REDIS_CACHE_ENV_CANADA=True
 REDIS_CACHE_NOAA=True
 # how long to cache data download from environment canada for.
 REDIS_ENV_CANADA_CACHE_EXPIRY=21600
+REDIS_NOAA_CACHE_EXPIRY=21600
 # c-haines tiff output is a feature that's useful for debugging - not intended to be set to true anywhere
 # other than on a developers machine.
 C_HAINES_OUTPUT_TIFF=False
